51-usb-device.rules
this file will allow mayflash dolphin bar to work as a mouse so that you can use your wiimote as a light gun. Will need to be install in udev/rules. Please note you do not setup the controls using es you will need to configure your controls in the mame menu press tab on a keyboard.